# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Incident:
    properties:
      assigned_to_team_id:
        type: string
      created_at:
        type: string
      created_by_user_id:
        type: string
      description:
        type: string
      id:
        type: string
      organization_id:
        type: string
      priority:
        $ref: '#/definitions/domain.AlertPriority'
      resolved_at:
        type: string
      severity:
        $ref: '#/definitions/domain.IncidentSeverity'
      started_at:
        type: string
      status:
        $ref: '#/definitions/domain.IncidentStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  domain.IncidentAlert:
    properties:
      alert_id:
        type: string
      id:
        type: string
      incident_id:
        type: string
      linked_at:
        type: string
      linked_by_user_id:
        type: string
    type: object
  domain.IncidentAlertWithDetails:
    properties:
      alert:
        $ref: '#/definitions/domain.Alert'
      alert_id:
        type: string
      id:
        type: string
      incident_id:
        type: string
      linked_at:
        type: string
      linked_by_user_id:
        type: string
    type: object
...
  service.CreateIncidentRequest:
    properties:
      assigned_to_team_id:
        type: string
      description:
        type: string
      priority:
        type: string
      severity:
        type: string
      title:
        type: string
    required:
    - priority
    - severity
    - title
    type: object
...
  service.LinkAlertRequest:
    properties:
      alert_id:
        type: string
    required:
    - alert_id
    type: object
...
  service.ListIncidentsResponse:
    properties:
      incidents:
        items:
          $ref: '#/definitions/domain.Incident'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  service.UpdateIncidentRequest:
    properties:
      assigned_to_team_id:
        type: string
      description:
        type: string
      priority:
        type: string
      severity:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pulsar.io
    name: API Support
    url: https://github.com/nmn3m/pulsar
  description: Pulsar is an incident management and alerting platform. This API provides
    endpoints for managing alerts, incidents, teams, schedules, escalation policies,
    notifications, and webhooks.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pulsar API
  version: "1.0"
paths:
...
  /incidents:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of incidents with optional filtering
        by status, severity, team, and search term
      parameters:
      - collectionFormat: csv
        description: Filter by status (investigating, identified, monitoring, resolved)
        in: query
        items:
          type: string
        name: status
        type: array
      - collectionFormat: csv
        description: Filter by severity (critical, high, medium, low)
        in: query
        items:
          type: string
        name: severity
        type: array
      - description: Filter by assigned team ID
        format: uuid
        in: query
        name: assigned_to_team_id
        type: string
      - description: Search term for title and description
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ListIncidentsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List incidents
      tags:
      - Incidents
    post:
      consumes:
      - application/json
      description: Creates a new incident with the provided details. The incident
        is created in the investigating status.
      parameters:
      - description: Incident creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateIncidentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Incident'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new incident
      tags:
      - Incidents
  /incidents/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes an incident and all associated data
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an incident
      tags:
      - Incidents
    get:
      consumes:
      - application/json
      description: Retrieves an incident with all related data including responders,
        alerts, and timeline events
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.IncidentWithDetails'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an incident with full details
      tags:
      - Incidents
    patch:
      consumes:
      - application/json
      description: Updates an existing incident with the provided fields. Only fields
        included in the request body will be updated.
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Incident update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateIncidentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Incident'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an incident
      tags:
      - Incidents
  /incidents/{id}/alerts:
    get:
      consumes:
      - application/json
      description: Retrieves all alerts that are associated with an incident
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.IncidentAlertWithDetails'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List alerts linked to an incident
      tags:
      - Incidents
    post:
      consumes:
      - application/json
      description: Associates an existing alert with an incident for tracking and
        correlation
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Link alert request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LinkAlertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IncidentAlert'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Link an alert to an incident
      tags:
      - Incidents
  /incidents/{id}/alerts/{alertId}:
    delete:
      consumes:
      - application/json
      description: Removes the association between an alert and an incident
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Alert ID
        format: uuid
        in: path
        name: alertId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unlink an alert from an incident
      tags:
      - Incidents
  /incidents/{id}/notes:
    post:
      consumes:
      - application/json
      description: Adds a note to the incident timeline as a timeline event
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Add note request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AddNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IncidentTimelineEvent'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a note to an incident
      tags:
      - Incidents
  /incidents/{id}/responders:
    get:
      consumes:
      - application/json
      description: Retrieves all responders assigned to an incident with their user
        details
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ResponderWithUser'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List responders for an incident
      tags:
      - Incidents
    post:
      consumes:
      - application/json
      description: Assigns a user as a responder to an incident with a specific role
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Add responder request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AddResponderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IncidentResponder'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a responder to an incident
      tags:
      - Incidents
  /incidents/{id}/responders/{responderId}:
    delete:
      consumes:
      - application/json
      description: Removes a user from the list of responders for an incident
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Responder ID
        format: uuid
        in: path
        name: responderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a responder from an incident
      tags:
      - Incidents
    patch:
      consumes:
      - application/json
      description: Updates the role of a responder assigned to an incident
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Responder ID
        format: uuid
        in: path
        name: responderId
        required: true
        type: string
      - description: Update responder role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateResponderRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a responder's role
      tags:
      - Incidents
  /incidents/{id}/timeline:
    get:
      consumes:
      - application/json
      description: Retrieves all timeline events for an incident including status
        changes, notes, and responder actions
      parameters:
      - description: Incident ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TimelineEventWithUser'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get incident timeline
      tags:
      - Incidents
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"